refactored role + categories config and tests
